# backend/internal/adapters/http/handlers/openapi/swagger.yaml
components:
  schemas:
    dto.AmountModel:
      description: '@Description Custom price for this subscription'
      properties:
        currency:
          example: USD
          type: string
        source:
          $ref: '#/components/schemas/dto.AmountModel'
        value:
          example: 100
          type: number
      required:
      - currency
      - value
      type: object
...
    dto.CreateSubscriptionRequest:
      properties:
        created_at:
          type: string
        custom_recurrency:
          type: integer
        end_date:
          format: date-time
          type: string
        family_users:
          items:
            type: string
          type: array
          uniqueItems: false
        free_trial:
          $ref: '#/components/schemas/dto.SubscriptionFreeTrialModel'
        friendly_name:
          type: string
        id:
          type: string
        labels:
          items:
            type: string
          type: array
          uniqueItems: false
        owner:
          enum:
          - personal
          - family
          - system
          example: personal
          type: string
        payer:
          $ref: '#/components/schemas/dto.EditableSubscriptionPayerModel'
        price:
          $ref: '#/components/schemas/dto.AmountModel'
        provider_id:
          type: string
        provider_key:
          type: string
        recurrency:
          type: string
        start_date:
          format: date-time
          type: string
      required:
      - owner
      - recurrency
      - start_date
      type: object
...
    dto.EditableSubscriptionPayerModel:
      description: Subscription payer object used for updating who pays for a subscription
      properties:
        memberId:
          description: '@Description LabelID of the specific family member who pays
            (required when type is family_member)'
          example: 123e4567-e89b-12d3-a456-426614174001
          type: string
        type:
          description: '@Description Type of payer (family or family member)'
          enum:
          - family
          - family_member
          example: family_member
          type: string
      required:
      - type
      type: object
...
    dto.LabelRefModel:
      properties:
        label_id:
          example: 123e4567-e89b-12d3-a456-426614174000
          type: string
        source:
          enum:
          - subscription
          - provider
          example: subscription
          type: string
      required:
      - label_id
      - source
      type: object
    dto.OwnerModel:
      description: '@Description Ownership information specifying whether this subscription
        belongs to a user or family'
      properties:
        etag:
          description: '@Description Entity tag for optimistic concurrency control'
          example: W/"123456789"
          type: string
        family_id:
          description: '@Description Family LabelID when an ownership type is family
            (required for family ownership)'
          example: 123e4567-e89b-12d3-a456-426614174000
          type: string
        type:
          description: '@Description Type of ownership (personal, family or system)'
          enum:
          - personal
          - family
          - system
          example: personal
          type: string
        userId:
          description: '@Description UserProfile LabelID when an ownership type is
            personal (required for personal ownership)'
          example: 123e4567-e89b-12d3-a456-426614174001
          type: string
      required:
      - etag
      - type
      type: object
...
    dto.PaginatedResponseModel-SubscriptionModel:
      properties:
        data:
          description: Data contains the list of items for the current page
          items:
            $ref: '#/components/schemas/dto.SubscriptionModel'
          type: array
          uniqueItems: false
        length:
          description: Length represents the number of items in the current page
          type: integer
        total:
          description: Total represents the total number of items available
          type: integer
      required:
      - data
      - length
      - total
      type: object
...
    dto.SubscriptionFreeTrialModel:
      description: '@Description Number of free trial days remaining (null if no trial
        or trial expired)'
      properties:
        end_date:
          format: date-time
          type: string
        start_date:
          format: date-time
          type: string
      required:
      - end_date
      - start_date
      type: object
    dto.SubscriptionModel:
      description: Subscription object containing all information about an active
        subscription including billing and usage details
      properties:
        created_at:
          description: '@Description ISO 8601 timestamp when the subscription was
            originally created'
          example: "2023-01-15T10:30:00Z"
          format: date-time
          type: string
        custom_recurrency:
          description: '@Description CustomRecurrency recurrency interval in days
            (required when recurrency is custom)'
          example: 90
          maximum: 3650
          minimum: 1
          type: integer
        end_date:
          description: '@Description ISO 8601 timestamp when the subscription expires
            (null for ongoing subscriptions)'
          example: "2024-01-01T00:00:00Z"
          format: date-time
          type: string
        etag:
          description: '@Description Entity tag used for optimistic concurrency control
            to prevent conflicting updates'
          example: W/"123456789"
          type: string
        family_users:
          description: '@Description List of family member IDs who use this service
            (for shared subscriptions)'
          example:
          - 123e4567-e89b-12d3-a456-426614174005
          - 123e4567-e89b-12d3-a456-426614174006
          items:
            type: string
          type: array
          uniqueItems: false
        free_trial:
          $ref: '#/components/schemas/dto.SubscriptionFreeTrialModel'
        friendly_name:
          description: '@Description Optional custom name for easy identification
            of the subscription'
          example: Netflix Family Account
          maxLength: 255
          type: string
        id:
          description: '@Description Unique identifier for the subscription (UUID
            format)'
          example: 123e4567-e89b-12d3-a456-426614174000
          type: string
        is_active:
          description: '@Description Indicates whether the subscription is currently
            active or not'
          example: true
          type: boolean
        label_refs:
          description: '@Description List of labels associated with this subscription'
          items:
            $ref: '#/components/schemas/dto.LabelRefModel'
          type: array
          uniqueItems: false
        owner:
          $ref: '#/components/schemas/dto.OwnerModel'
        payer:
          $ref: '#/components/schemas/dto.SubscriptionPayerModel'
        price:
          $ref: '#/components/schemas/dto.AmountModel'
        provider_id:
          description: '@Description LabelID of the service provider offering this
            subscription'
          example: 123e4567-e89b-12d3-a456-426614174002
          type: string
        recurrency:
          description: '@Description Billing recurrency pattern (monthly, yearly,
            custom, etc.)'
          enum:
          - unknown
          - one_time
          - monthly
          - quarterly
          - half_yearly
          - yearly
          - custom
          example: monthly
          type: string
        start_date:
          description: '@Description ISO 8601 timestamp when the subscription becomes
            active'
          example: "2023-01-01T00:00:00Z"
          format: date-time
          type: string
        updated_at:
          description: '@Description ISO 8601 timestamp when the subscription was
            last modified'
          example: "2023-01-20T14:45:30Z"
          format: date-time
          type: string
      required:
      - created_at
      - etag
      - id
      - is_active
      - owner
      - provider_id
      - recurrency
      - start_date
      - updated_at
      type: object
    dto.SubscriptionPayerModel:
      description: '@Description Information about who pays for this subscription
        within the family'
      properties:
        etag:
          description: '@Description Entity tag used for optimistic concurrency control
            to prevent conflicting updates'
          example: W/"123456789"
          type: string
        memberId:
          description: '@Description LabelID of the specific family member who pays
            (required when type is family_member)'
          example: 123e4567-e89b-12d3-a456-426614174001
          type: string
        type:
          description: '@Description Type of payer (family or family member)'
          enum:
          - family
          - family_member
          example: family_member
          type: string
      required:
      - etag
      - type
      type: object
...
    ginx.HttpErrorResponse:
      description: RFC7807 Problem Details error response
      properties:
        detail:
          example: Missing required field 'name'
          type: string
        instance:
          example: /api/resource/123
          type: string
        status:
          example: 400
          type: integer
        title:
          example: Bad Request
          type: string
        type:
          example: about:blank
          type: string
      type: object
...
info:
  contact:
    email: support@mistribe.com
    name: API Support
    url: http://subtracker.mistribe.com/support
  description: This api provide HTTPRest endpoints for the application SubTracker.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://subtracker.mistribe.com/terms/
  title: SubTracker API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /subscriptions:
    get:
      description: Retrieve a paginated list of all subscriptions for the authenticated
        user
      parameters:
      - description: Search text
        in: query
        name: search
        schema:
          type: string
      - description: Filter by recurrency types
        in: query
        name: recurrencies
        schema:
          items:
            type: string
          type: array
      - description: Filter by start date (RFC3339)
        in: query
        name: from_date
        schema:
          type: string
      - description: Filter by end date (RFC3339)
        in: query
        name: to_date
        schema:
          type: string
      - description: Filter by user IDs
        in: query
        name: users
        schema:
          items:
            type: string
          type: array
      - description: Include inactive subscriptions
        in: query
        name: with_inactive
        schema:
          type: boolean
      - description: Filter by provider IDs
        in: query
        name: providers
        schema:
          items:
            type: string
          type: array
      - description: 'Number of items per page (default: 10)'
        in: query
        name: limit
        schema:
          type: integer
      - description: 'Page number (default: 0)'
        in: query
        name: offset
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.PaginatedResponseModel-SubscriptionModel'
          description: Paginated list of subscriptions
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Bad Request - Invalid query parameters
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Internal Server Error
      summary: Get all subscriptions
      tags:
      - subscriptions
    post:
      description: Create a new subscription with provider, plan, pricing, and payment
        information
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.CreateSubscriptionRequest'
        description: Subscription creation data
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.SubscriptionModel'
          description: Successfully created subscription
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Bad Request - Invalid input data
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Unauthorized - Invalid user authentication
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Internal Server Error
      summary: Create a new subscription
      tags:
      - subscriptions
...
  /subscriptions/export:
    get:
      description: Export all subscriptions in CSV, JSON, or YAML format
      parameters:
      - description: Export format (csv, json, yaml)
        in: query
        name: format
        schema:
          default: json
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: file
            application/x-yaml:
              schema:
                type: string
          description: Exported subscriptions file
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Invalid format parameter
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ginx.HttpErrorResponse'
          description: Internal Server Error
      summary: Export subscriptions
      tags:
      - subscriptions
...
servers:
- description: Production server
  url: https://api.subtracker.mistribe.com
- description: Development server
  url: http://localhost:8080